+++ eras/allegra/impl/CHANGELOG.md
@@ -29,6 +29,8 @@
 * Add `ToJSON` and `FromJSON` instances for
   - `AllegraTxAuxData era`
   - `Timelock era`
+  - `TxBody TopTx AllegraEra`
+  - `Tx TopTx AllegraEra`
 * Export `allegraBasedEraNativeScriptToJSON` and `allegraBasedEraNativeScriptJSONParser` from `Cardano.Ledger.Allegra.Scripts`
 
 ### `cddl`